requirements.txt
requests==2.31.0
textblob==0.18.0
vaderSentiment==3.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.35.0
pandas==2.1.4
plotly==5.22.0
